Move whoami module to auth directory
Also use `GlobusApp` instead of `LoginManager`
